@@ -39,29 +39,28 @@
 		"format": "prettier --plugin-search-dir . --write .",
 		"test": "tsd && jest test --coverage"
 	},
+	"dependencies": {
+		"@sveltejs/kit": "^1.15.7",
+		"svelte": "^5.0.0-next"
+	},
 	"devDependencies": {
-		"@fortawesome/fontawesome-common-types": "^6.2.0",
-		"@sveltejs/adapter-auto": "next",
-		"@sveltejs/kit": "next",
-		"@sveltejs/package": "next",
-		"@typescript-eslint/eslint-plugin": "^5.42.1",
-		"@typescript-eslint/parser": "^5.42.1",
-		"eslint": "^8.27.0",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-jest": "^27.1.5",
+		"@fortawesome/fontawesome-common-types": "^6.4.0",
+		"@fortawesome/free-solid-svg-icons": "^6.4.0",
+		"@sveltejs/adapter-auto": "^2.0.0",
+		"@sveltejs/package": "^2.0.2",
+		"@typescript-eslint/eslint-plugin": "^5.59.0",
+		"@typescript-eslint/parser": "^5.59.0",
+		"eslint": "^8.39.0",
+		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-svelte3": "^4.0.0",
-		"jest": "^29.3.1",
-		"jest-environment-jsdom": "^29.3.1",
-		"prettier": "^2.7.1",
-		"prettier-plugin-svelte": "^2.8.0",
-		"svelte": "^3.53.1",
-		"svelte-check": "^2.9.2",
-		"svelte-jester": "^2.3.2",
-		"svelte-preprocess": "^4.10.7",
-		"tsd": "^0.24.1",
-		"tslib": "^2.4.1",
-		"typescript": "^4.8.4",
-		"vite": "^3.2.3"
+		"prettier": "^2.8.8",
+		"prettier-plugin-svelte": "^2.10.0",
+		"svelte-check": "^3.2.0",
+		"svelte-preprocess": "^5.0.3",
+		"tsd": "^0.28.1",
+		"tslib": "^2.5.0",
+		"typescript": "^5.0.4",
+		"vite": "^4.3.1"
 	},
 	"type": "module",
 	"exports": {
